api/pyproject.toml
[tool.poetry]
name = "ai-teacher-api-server"
version = "0.1.0"
description = ""
authors = ["amazingchow <jianzhou42@163.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
celery = "^5.4.0"
sqlalchemy = "2.0.19"
fastapi = "^0.115.11"
python-levenshtein = "^0.27.1"
openai-whisper = "20240930"
simhash = "2.1.2"
uvicorn = "^0.34.0"
aiosqlite = "0.19.0"
greenlet = "3.1.1"
python-multipart = "0.0.6"
opencc-python-reimplemented = "^0.1.7"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.doubanio.com/simple/"
priority = "primary"

[tool.sort.config]
auto = true
case-sensitive = false
sort-python = false
format = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "asyncio"
]
[tool.isort]
# More info: https://pycqa.github.io/isort/docs/configuration/action_comments.html
profile = "hug"
src_paths = ["."]
known_local_folder = [
    "celery_app",
    "database",
    "models",
    "schemas",
    "tasks"
]
sections = ["FUTURE", "STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]